+++ user/src/main/java/com/team15gijo/user/application/service/UserApplicationService.java
@@ -4,14 +4,14 @@ import com.team15gijo.user.domain.model.UserStatus;
 import com.team15gijo.user.infrastructure.dto.UserFeignInfoResponseDto;
 import com.team15gijo.user.presentation.dto.v1.AdminUserReadResponseDto;
 import com.team15gijo.user.presentation.dto.v1.UserReadsResponseDto;
-import com.team15gijo.user.presentation.dto.v1.request.AdminUserStatusUpdateRequestDto;
-import com.team15gijo.user.presentation.dto.v1.request.UserEmailUpdateRequestDto;
-import com.team15gijo.user.presentation.dto.v1.request.UserPasswordUpdateRequestDto;
-import com.team15gijo.user.presentation.dto.v1.request.UserSignUpRequestDto;
-import com.team15gijo.user.presentation.dto.v1.request.UserUpdateRequestDto;
-import com.team15gijo.user.presentation.dto.v1.response.UserReadResponseDto;
-import com.team15gijo.user.presentation.dto.v1.response.UserSignUpResponseDto;
-import com.team15gijo.user.presentation.dto.v1.response.UserUpdateResponseDto;
+import com.team15gijo.user.presentation.dto.request.v1.AdminUserStatusUpdateRequestDto;
+import com.team15gijo.user.presentation.dto.request.v1.UserEmailUpdateRequestDto;
+import com.team15gijo.user.presentation.dto.request.v1.UserPasswordUpdateRequestDto;
+import com.team15gijo.user.presentation.dto.request.v1.UserSignUpRequestDto;
+import com.team15gijo.user.presentation.dto.request.v1.UserUpdateRequestDto;
+import com.team15gijo.user.presentation.dto.response.v1.UserReadResponseDto;
+import com.team15gijo.user.presentation.dto.response.v1.UserSignUpResponseDto;
+import com.team15gijo.user.presentation.dto.response.v1.UserUpdateResponseDto;
 import jakarta.validation.Valid;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
