+++ user/src/main/java/com/team15gijo/user/application/service/v1/UserApplicationServiceImpl.java
@@ -11,20 +11,20 @@ import com.team15gijo.user.domain.repository.UserRepository;
 import com.team15gijo.user.domain.service.UserDomainService;
 import com.team15gijo.user.infrastructure.client.AuthServiceClient;
 import com.team15gijo.user.infrastructure.dto.UserFeignInfoResponseDto;
-import com.team15gijo.user.infrastructure.dto.v1.internal.AuthSignUpRequestDto;
-import com.team15gijo.user.infrastructure.dto.v1.internal.AuthSignUpUpdateUserIdRequestDto;
+import com.team15gijo.user.infrastructure.dto.request.v1.AuthSignUpRequestDto;
+import com.team15gijo.user.infrastructure.dto.request.v1.AuthSignUpUpdateUserIdRequestDto;
 import com.team15gijo.user.presentation.dto.v1.AdminUserReadResponseDto;
 import com.team15gijo.user.presentation.dto.v1.UserReadsResponseDto;
-import com.team15gijo.user.presentation.dto.v1.internal.request.AuthIdentifierUpdateRequestDto;
-import com.team15gijo.user.presentation.dto.v1.internal.request.AuthPasswordUpdateRequestDto;
-import com.team15gijo.user.presentation.dto.v1.request.AdminUserStatusUpdateRequestDto;
-import com.team15gijo.user.presentation.dto.v1.request.UserEmailUpdateRequestDto;
-import com.team15gijo.user.presentation.dto.v1.request.UserPasswordUpdateRequestDto;
-import com.team15gijo.user.presentation.dto.v1.request.UserSignUpRequestDto;
-import com.team15gijo.user.presentation.dto.v1.request.UserUpdateRequestDto;
-import com.team15gijo.user.presentation.dto.v1.response.UserReadResponseDto;
-import com.team15gijo.user.presentation.dto.v1.response.UserSignUpResponseDto;
-import com.team15gijo.user.presentation.dto.v1.response.UserUpdateResponseDto;
+import com.team15gijo.user.infrastructure.dto.request.v1.AuthIdentifierUpdateRequestDto;
+import com.team15gijo.user.infrastructure.dto.request.v1.AuthPasswordUpdateRequestDto;
+import com.team15gijo.user.presentation.dto.request.v1.AdminUserStatusUpdateRequestDto;
+import com.team15gijo.user.presentation.dto.request.v1.UserEmailUpdateRequestDto;
+import com.team15gijo.user.presentation.dto.request.v1.UserPasswordUpdateRequestDto;
+import com.team15gijo.user.presentation.dto.request.v1.UserSignUpRequestDto;
+import com.team15gijo.user.presentation.dto.request.v1.UserUpdateRequestDto;
+import com.team15gijo.user.presentation.dto.response.v1.UserReadResponseDto;
+import com.team15gijo.user.presentation.dto.response.v1.UserSignUpResponseDto;
+import com.team15gijo.user.presentation.dto.response.v1.UserUpdateResponseDto;
 import java.util.UUID;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
