+++ user/src/main/java/com/team15gijo/user/domain/service/UserDomainService.java
@@ -1,7 +1,7 @@
 package com.team15gijo.user.domain.service;
 
 import com.team15gijo.user.domain.model.UserEntity;
-import com.team15gijo.user.presentation.dto.v1.request.UserSignUpRequestDto;
+import com.team15gijo.user.presentation.dto.request.v1.UserSignUpRequestDto;
 
 public interface UserDomainService {
