--- user/src/main/java/com/team15gijo/user/infrastructure/dto/v1/internal/UserSearchResponseDto.java
@@ -1,30 +1,0 @@
-package com.team15gijo.user.infrastructure.dto.v1.internal;
-
-import com.team15gijo.user.domain.model.UserEntity;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class UserSearchResponseDto {
-
-    private Long userId;
-    private String username;
-    private String nickname;
-    private String profile;
-    private String region;
-
-    public static UserSearchResponseDto from(UserEntity user) {
-        return UserSearchResponseDto.builder()
-                .userId(user.getId())
-                .username(user.getUsername())
-                .nickname(user.getNickname())
-                .profile(user.getProfile())
-                .region(user.getRegion())
-                .build();
-    }
-}
