+++ user/src/main/java/com/team15gijo/user/presentation/controller/v1/UserInternalController.java
@@ -4,7 +4,7 @@ import com.team15gijo.common.dto.ApiResponse;
 import com.team15gijo.user.application.service.UserApplicationService;
 import com.team15gijo.user.application.service.v1.InternalUserService;
 import com.team15gijo.user.infrastructure.dto.UserFeignInfoResponseDto;
-import com.team15gijo.user.infrastructure.dto.v1.internal.UserSearchResponseDto;
+import com.team15gijo.user.presentation.dto.internal.response.v1.UserSearchResponseDto;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
