+++ user/src/main/java/com/team15gijo/user/presentation/dto/request/v1/UserUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.team15gijo.user.presentation.dto.v1.request;
+package com.team15gijo.user.presentation.dto.request.v1;
 
 public record UserUpdateRequestDto(
         String username,
